Dijkstra with working checkpoints
For now, the path is not the optimal. Indeed it take the order of placement of the checkpoint and not the optimal order. Then, the path between two point is optimal, but not necessarily the complete path.

Maybe an upgrade would be :
Allowing the user to choose which method to use. Either by creating a check box or a new instance of algorithm